+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # CHANGELOG
 
+## v5.25.0 - 2021-11-21
+
+### Added
+
+- New tech cards:
+  - Sub Commander Duplication Tech
+  - Sub Commander Fabber Tech
+  - Sub Commander Tactics Tech
+
+### Fixed
+
+- Inventory tooltips clashing with system tooltips when four techs were displayed
+
 ## v5.24.2 - 2021-11-19
 
 ### Fixed